+++ tpie/sort_manager.h
@@ -451,7 +451,7 @@ err sort_manager<T,I,M>::compute_sort_params(void){
 	// number of substreams we want.  It may not be able to due to
 	// operating system restrictions, such as on the number of regions
 	// that can be mmap()ed in, max number of file descriptors, etc.
-	int availableStreams = inStream->available_streams ();
+	int availableStreams = static_cast<int>(inStream->available_streams ());
 	    
 	// Merging requires an available stream/file decriptor for
 	// each of the mrgArity input strems. We need one additional file descriptor
